backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pyaudio==0.2.14
requests==2.31.0
psutil==5.9.6
cryptography==41.0.7

# OVOS Wake Word Detection
ovos-plugin-manager==1.0.3
ovos-ww-plugin-vosk
sounddevice
numpy
SpeechRecognition
